UI code implemented

+++ Assets/Scripts/Core/CurrencyExchanger.cs
@@ -13,12 +13,12 @@ namespace CurrencyExchanger
             get;
             private set;
         }
-        public void Convert(string currencyFrom, string currencyTo, int amount, Action<float> exhangeRetrievedCallback, Action<string> errorCallback)
+        public void Convert(string currencyFrom, string currencyTo, float amount, Action<float> exhangeRetrievedCallback, Action<string> errorCallback)
         {
             backEndProvider.GetExchangeRateFor(currencyFrom, currencyTo, (exchangeRate) => exhangeRetrievedCallback(exchangeRate * amount), errorCallback);
         }
 
-        public void Convert(string currencyFrom, string currencyTo, int amount, DateTime date, Action<float> exhangeRetrievedCallback, Action<string> errorCallback)
+        public void Convert(string currencyFrom, string currencyTo, float amount, DateTime date, Action<float> exhangeRetrievedCallback, Action<string> errorCallback)
         {
             backEndProvider.GetExchangeRateFor(currencyFrom, currencyTo, date, (exchangeRate) => exhangeRetrievedCallback(exchangeRate * amount), errorCallback);
         }
